+++ src/ClassicUO.Client/LegionScripting/docs/API.py
@@ -40,12 +40,13 @@ class ApiGameObject:
     IsDestroyed: bool = None
     __class__: str = None
 
-    def SetOutlineColor(self, htmlColor: "str") -> None:
+    def SetOutlineColor(self, htmlColor: "str" = None) -> None:
         """
-         Set an objects outline color using html hex colors.
+         Set an object's outline color using HTML hex colors.
          Example:
          ```py
          API.Player.SetOutlineColor("#105510")
+         API.Player.SetOutlineColor(None)
          ```
         
         """